application/requirements.txt
bottle==0.12.16
psycopg2-binary==2.7.7
